🎨 callToastNotify를 Util 함수화 #11
